commented dismiss and separated module
