+++ apps/forms-flow-ai/forms-flow-web/src/actions/FOI/foiNotificationActions.js
@@ -1,0 +1,7 @@
+import FOI_ACTION_CONSTANTS from './foiActionConstants'
+export const setFOINotifications = (data) => dispatch =>{
+    dispatch({
+        type:FOI_ACTION_CONSTANTS.FOI_NOTIFICATIONS,
+        payload:data
+    })
+}
