+++ apps/forms-flow-ai/forms-flow-web/src/actions/FOI/foiRequestActions.js
@@ -176,9 +176,3 @@ export const setRequestAttachments = (data) => dispatch => {
     })
 }
 
-export const setFOINotifications = (data) => dispatch =>{
-    dispatch({
-        type:FOI_ACTION_CONSTANTS.FOI_NOTIFICATIONS,
-        payload:data
-    })
-}
