+++ apps/forms-flow-ai/forms-flow-web/src/apiManager/httpRequestHandler/index.js
@@ -64,7 +64,3 @@ export const httpDELETERequest = (url, data, token, isBearer = true) => {
   });
 };
 
-/*export const httpPUTRequest = (url, data, token, isBearer=true) => {
-  return axios.put(url, data, { headers: { Authorization: isBearer ?`Bearer ${ token || UserService.getToken()}`: token } });
-};*/
-
