+++ apps/forms-flow-ai/forms-flow-web/src/apiManager/services/FOI/foiRequestServices.js
@@ -1,7 +1,6 @@
 import {
   httpPOSTRequest,
   httpGETRequest,
-  httpDELETERequest,
 } from "../../httpRequestHandler";
 import API from "../../endpoints";
 import {
@@ -13,7 +12,6 @@ import {
   clearRequestDetails,
   setFOIRequestDescriptionHistory,
   setFOIMinistryRequestList,
-  setFOINotifications,
 } from "../../../actions/FOI/foiRequestActions";
 import { fetchFOIAssignedToList, fetchFOIMinistryAssignedToList } from "./foiMasterDataServices";
 import { catchError, fnDone } from './foiServicesUtil';
@@ -276,54 +274,3 @@ export const fetchFOIRequestDescriptionList = (requestId, ministryId) => {
   };
 };
 
-  export const fetchFOINotifications = (...rest) => {
-    const done = rest.length ? rest[0] : () => { };
-    return (dispatch) => {
-      httpGETRequest(API.FOI_GET_NOTIFICATIONS, {}, UserService.getToken())
-        .then((res) => {
-          if (res.data) {
-            let data = res.data.map((foiNotifications) => {
-              return { ...foiNotifications };
-            });
-            dispatch(setFOINotifications(data));
-            dispatch(setFOILoader(false));
-            done(null, res.data);
-          } else {
-            console.log("Error", res);
-            dispatch(serviceActionError(res));
-            dispatch(setFOILoader(false));
-          }
-        })
-        .catch((error) => {
-          console.log("Error", error);
-          dispatch(serviceActionError(error));
-          dispatch(setFOILoader(false));
-          done(error);
-        });
-    };
-  };
-
-  export const deleteFOINotification = (idNumber, notificationId,data, ...rest) => {
-    const done = rest.length ? rest[0] : () => { };
-    let apiUrl = replaceUrl(replaceUrl(
-      API.FOI_DELETE_NOTIFICATION,
-      "<idNumber>",
-      idNumber
-    ), "<notificationId>", notificationId);
-    return (dispatch) => {
-      httpDELETERequest(apiUrl, data)
-        .then((res) => {
-          dispatch(fetchFOINotifications())
-          if (res.data) {
-            done(null, res.data);
-          } else {
-            dispatch(serviceActionError(res));
-            done("Error dismissing Notification");
-          }
-        })
-        .catch((error) => {
-          dispatch(serviceActionError(error));
-          done(error);
-        });
-    };
-  };
