+++ apps/forms-flow-ai/forms-flow-web/src/components/FOI/Header/FOIHeader.js
@@ -11,7 +11,7 @@ import Popup from 'reactjs-popup';
 import NotificationPopup from "./NotificationPopup/NotificationPopup";
 import {
   fetchFOINotifications
-} from "../../../apiManager/services/FOI/foiRequestServices";
+} from "../../../apiManager/services/FOI/foiNotificationServices";
 
 const FOIHeader = React.memo(() => { 
 
@@ -33,10 +33,11 @@ const openModal = (coordinates) => {
   setOpen(!open);
 }
 
-let foiNotifications = useSelector(state=> state.foiRequests.foiNotifications);
+let foiNotifications = useSelector(state=> state.notifications.foiNotifications);
 
 useEffect(() => {     
-  dispatch(fetchFOINotifications());
+  if(isAuthenticated)
+    dispatch(fetchFOINotifications());
   const fetchNotifications = () =>{
     dispatch(fetchFOINotifications());
   }
