+++ apps/forms-flow-ai/forms-flow-web/src/components/FOI/Header/NotificationPopup/NotificationList/notificationlist.scss
@@ -28,10 +28,6 @@
     margin-bottom: 10px;
 }
 
-.redirect-url h1 {
-    color: #000000;
-}
-
 .notification-heading > a {
     text-decoration: none!important;
     color: #000000;
