+++ apps/forms-flow-ai/forms-flow-web/src/modules/FOI/foiRequestsReducer.js
@@ -22,7 +22,6 @@ const initialState = {
   foiWatcherList: [],
   foiRequestComments:[],
   foiRequestAttachments:[],
-  foiNotifications: []
 }
 
 const foiRequests = (state = initialState, action)=> {
@@ -75,8 +74,6 @@ const foiRequests = (state = initialState, action)=> {
       return {...state, foiRequestComments: action.payload};
     case FOI_ACTION_CONSTANTS.FOI_REQUEST_ATTACHMENTS:
       return {...state, foiRequestAttachments: action.payload};
-    case FOI_ACTION_CONSTANTS.FOI_NOTIFICATIONS:
-      return {...state, foiNotifications: action.payload};   
     default:
       return state;
   }
